+++ deploy/glider.sql
@@ -6,9 +6,9 @@
 BEGIN;
 
 CREATE TABLE gwapese.glider (
-  glider_id smallint NOT NULL,
+  glider_id integer NOT NULL,
   icon text NOT NULL,
-  presentation_order smallint NOT NULL,
+  presentation_order integer NOT NULL,
   CONSTRAINT glider_pk PRIMARY KEY (glider_id)
 );
 
@@ -24,7 +24,7 @@ CALL temporal_tables.create_historicize_trigger ('gwapese',
 CREATE TABLE gwapese.glider_description (
   app_name text NOT NULL,
   lang_tag text NOT NULL,
-  glider_id smallint NOT NULL,
+  glider_id integer NOT NULL,
   original text NOT NULL,
   CONSTRAINT glider_description_pk PRIMARY KEY (app_name, lang_tag, glider_id),
   CONSTRAINT glider_identifies_glider_description_fk FOREIGN KEY (glider_id)
@@ -44,9 +44,9 @@ CALL temporal_tables.create_historicize_trigger ('gwapese',
   'glider_description', 'glider_description_history');
 
 CREATE TABLE gwapese.glider_dye_slot (
-  color_id smallint NOT NULL,
-  glider_id smallint NOT NULL,
-  slot_index smallint NOT NULL,
+  color_id integer NOT NULL,
+  glider_id integer NOT NULL,
+  slot_index integer NOT NULL,
   CONSTRAINT glider_dye_slot_pk PRIMARY KEY (glider_id, slot_index),
   CONSTRAINT color_identifies_glider_dye_slot_fk FOREIGN KEY (color_id)
     REFERENCES gwapese.color (color_id) ON DELETE CASCADE ON UPDATE CASCADE,
@@ -66,7 +66,7 @@ CALL temporal_tables.create_historicize_trigger ('gwapese',
 CREATE TABLE gwapese.glider_name (
   app_name text NOT NULL,
   lang_tag text NOT NULL,
-  glider_id smallint NOT NULL,
+  glider_id integer NOT NULL,
   original text NOT NULL,
   CONSTRAINT glider_name_pk PRIMARY KEY (app_name, lang_tag, glider_id),
   CONSTRAINT glider_identifies_glider_name_fk FOREIGN KEY (glider_id)
