+++ deploy/skiff.sql
@@ -7,7 +7,7 @@ BEGIN;
 
 CREATE TABLE gwapese.skiff (
   icon text NOT NULL,
-  skiff_id smallint NOT NULL,
+  skiff_id integer NOT NULL,
   CONSTRAINT skiff_pk PRIMARY KEY (skiff_id)
 );
 
@@ -21,10 +21,10 @@ CALL temporal_tables.create_historicize_trigger ('gwapese',
   'skiff', 'skiff_history');
 
 CREATE TABLE gwapese.skiff_dye_slot (
-  color_id smallint NOT NULL,
+  color_id integer NOT NULL,
   material text NOT NULL,
-  skiff_id smallint NOT NULL,
-  slot_index smallint NOT NULL,
+  skiff_id integer NOT NULL,
+  slot_index integer NOT NULL,
   CONSTRAINT skiff_dye_slot_pk PRIMARY KEY (skiff_id, slot_index),
   CONSTRAINT skiff_contains_skiff_dye_slot_fk FOREIGN KEY (skiff_id) REFERENCES
     gwapese.skiff (skiff_id) ON DELETE CASCADE ON UPDATE CASCADE,
@@ -46,7 +46,7 @@ CREATE TABLE gwapese.skiff_name (
   app_name text NOT NULL,
   lang_tag text NOT NULL,
   original text NOT NULL,
-  skiff_id smallint NOT NULL,
+  skiff_id integer NOT NULL,
   CONSTRAINT skiff_name_pk PRIMARY KEY (app_name, lang_tag, skiff_id),
   CONSTRAINT skiff_identifies_skiff_name_fk FOREIGN KEY (skiff_id) REFERENCES
     gwapese.skiff (skiff_id) ON DELETE CASCADE ON UPDATE CASCADE,
