+++ deploy/skin.sql
@@ -7,7 +7,7 @@ BEGIN;
 CREATE TABLE gwapese.skin (
   icon text,
   rarity text NOT NULL,
-  skin_id smallint NOT NULL,
+  skin_id integer NOT NULL,
   CONSTRAINT skin_pk PRIMARY KEY (skin_id)
 );
 
@@ -24,7 +24,7 @@ CREATE TABLE gwapese.skin_description (
   app_name text NOT NULL,
   lang_tag text NOT NULL,
   original text NOT NULL,
-  skin_id smallint NOT NULL,
+  skin_id integer NOT NULL,
   CONSTRAINT skin_description_pk PRIMARY KEY (app_name, lang_tag, skin_id),
   CONSTRAINT operating_copy_precedes_skin_description_fk FOREIGN KEY (app_name,
     lang_tag, original) REFERENCES gwapese.operating_copy (app_name, lang_tag,
@@ -44,7 +44,7 @@ CALL temporal_tables.create_historicize_trigger ('gwapese',
 
 CREATE TABLE gwapese.skin_flag (
   flag text NOT NULL,
-  skin_id smallint NOT NULL,
+  skin_id integer NOT NULL,
   CONSTRAINT skin_flag_pk PRIMARY KEY (skin_id, flag),
   CONSTRAINT skin_identifies_skin_flag_fk FOREIGN KEY (skin_id) REFERENCES
     gwapese.skin (skin_id) ON DELETE CASCADE ON UPDATE CASCADE
@@ -63,7 +63,7 @@ CREATE TABLE gwapese.skin_name (
   app_name text NOT NULL,
   original text NOT NULL,
   lang_tag text NOT NULL,
-  skin_id smallint NOT NULL,
+  skin_id integer NOT NULL,
   CONSTRAINT skin_name_pk PRIMARY KEY (app_name, lang_tag, skin_id),
   CONSTRAINT operating_copy_precedes_skin_name_fk FOREIGN KEY (app_name,
     lang_tag, original) REFERENCES gwapese.operating_copy (app_name, lang_tag,
@@ -83,7 +83,7 @@ CALL temporal_tables.create_historicize_trigger ('gwapese',
 
 CREATE TABLE gwapese.skin_restriction (
   restriction text NOT NULL,
-  skin_id smallint NOT NULL,
+  skin_id integer NOT NULL,
   CONSTRAINT skin_restriction_pk PRIMARY KEY (skin_id, restriction),
   CONSTRAINT skin_identifies_skin_restriction_fk FOREIGN KEY (skin_id)
     REFERENCES gwapese.skin (skin_id) ON DELETE CASCADE,
@@ -101,7 +101,7 @@ CALL temporal_tables.create_historicize_trigger ('gwapese',
   'skin_restriction', 'skin_restriction_history');
 
 CREATE TABLE gwapese.skin_type (
-  skin_id smallint UNIQUE NOT NULL,
+  skin_id integer UNIQUE NOT NULL,
   skin_type text NOT NULL,
   CONSTRAINT skin_type_pk PRIMARY KEY (skin_id, skin_type),
   CONSTRAINT skin_identifies_skin_type_fk FOREIGN KEY (skin_id) REFERENCES
