+++ deploy/skin_weapon.sql
@@ -8,7 +8,7 @@ BEGIN;
 
 CREATE TABLE gwapese.skin_weapon (
   damage_type text NOT NULL,
-  skin_id smallint UNIQUE NOT NULL,
+  skin_id integer UNIQUE NOT NULL,
   skin_type text GENERATED ALWAYS AS ('Weapon') STORED,
   weapon_type text NOT NULL,
   CONSTRAINT skin_weapon_pk PRIMARY KEY (skin_id, skin_type),
